Overall improvment on the dsl package.
The code is now more simple to understand (in the dsl package).
Add a lot of tests to ensure the dsl only allows the right nesting and
semantics. It is still necessary to test that a dsl.Expect generates the
correct core.Expect.
Factor some methods in core.Whens thus removing duplicate code.