Reasonably working new major version with read from semantics.
Whens now allow to specify from which InputStream the input will be read and matched.
Expect, ExpectBlock, When, Action are not covariant.
Update dependencies.
Update to sbt 1.0.0.
Settings is now a truly domain class.
Extract a trait for RichProcess to allow other implementations.
Tests for many of the new features not yet created.